Update the data collator for Luke
