+++ src/transformers/data/data_collator.py
@@ -70,6 +70,42 @@ def default_data_collator(features: List[InputDataClass], return_tensors="pt") -
         return numpy_default_data_collator(features)
 
 
+def luke_entity_span_classification_data_collator(batch, tokenizer):
+    import torch
+
+    def create_padded_sequence(target, padding_value):
+        if isinstance(target, str):
+            tensors = [torch.tensor(o[target], dtype=torch.long) for o in batch]
+        else:
+            tensors = [torch.tensor(o, dtype=torch.long) for o in target]
+        return torch.nn.utils.rnn.pad_sequence(tensors, batch_first=True, padding_value=padding_value)
+    
+    def padding_tensor(sequences):
+        num = len(sequences)
+        out_dims = (num, tokenizer.max_entity_length)
+        out_tensor = sequences[0].data.new(*out_dims).fill_(0)
+        for i, tensor in enumerate(sequences):
+            length = tensor.size(0)
+            out_tensor[i, :length] = tensor
+            
+        return out_tensor
+
+    ret = dict(
+        input_ids=create_padded_sequence("input_ids", tokenizer.pad_token_id),
+        attention_mask=create_padded_sequence("attention_mask", 0),
+        entity_start_positions=create_padded_sequence("entity_start_positions", 0),
+        entity_end_positions=create_padded_sequence("entity_end_positions", 0),
+        entity_ids=create_padded_sequence("entity_ids", 0),
+        entity_attention_mask=create_padded_sequence("entity_attention_mask", 0),
+        entity_position_ids=create_padded_sequence("entity_position_ids", -1),
+    )
+    ret["labels"] = padding_tensor(create_padded_sequence("labels", -100))
+    ret["original_entity_spans"] = [torch.tensor(o["original_entity_spans"], dtype=torch.long) for o in batch]
+    ret["ner_tags"] = [torch.tensor(o["ner_tags"], dtype=torch.long) for o in batch]
+
+    return ret
+
+
 @dataclass
 class DefaultDataCollator(DataCollatorMixin):
     """
